parallel processes in paper_experiment
